+++ dice/apps/games/admin.py
@@ -1,7 +1,13 @@
 from django.contrib import admin
-from .models import Room
+
+from dice.apps.games.models import Room, Game
 
 
 @admin.register(Room)
+class RoomAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(Game)
 class RoomAdmin(admin.ModelAdmin):
     pass
