+++ dice/apps/games/tests/factories.py
@@ -1,6 +1,5 @@
-import factory
 import factory.fuzzy
-from dice.apps.games.models import Room, Round, Game, Dice
+from dice.apps.games.models import Room, Game
 from dice.apps.users.tests.factories import UserFactory
 
 
@@ -12,23 +11,8 @@ class RoomFactory(factory.django.DjangoModelFactory):
         model = Room
 
 
-class DiceFactory(factory.django.DjangoModelFactory):
-    value = factory.fuzzy.FuzzyInteger(1, 6)
-
-    class Meta:
-        model = Dice
-
-
 class GameFactory(factory.django.DjangoModelFactory):
     room = factory.SubFactory(RoomFactory)
 
     class Meta:
         model = Game
-
-
-class RoundFactory(factory.django.DjangoModelFactory):
-    game = factory.SubFactory(GameFactory)
-    user = factory.SelfAttribute('game.room.host')
-
-    class Meta:
-        model = Round
