+++ dice/apps/rounds/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class RoundsConfig(AppConfig):
+    name = 'dice.apps.rounds'
