+++ dice/apps/rounds/tests/test_utilities.py
@@ -1,0 +1,2 @@
+# todo: test roll_dices
+
