+++ dice/apps/users/serializers.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
-from dice.apps.users.models import User
 from django.contrib.auth.hashers import make_password
+
+from dice.apps.users.models import User
 
 
 class UserSerializer(serializers.ModelSerializer):
