+++ dice/apps/users/tests/test_views.py
@@ -1,0 +1,3 @@
+# todo: testy do widoku CustomObtainAuthToken.post:
+#    a) token istnieje i test struktury response
+#    b) token nie istnieje i test response status code
