+++ dice/apps/users/tests/tests_serializers.py
@@ -1,0 +1,1 @@
+# todo: test UserSerializer: metoda create - test sprawdzajacy czy haslo zostalo stworzone (check_password)
